Big project restructuration
I modified the entire source tree and fixed the includes, this was made
in order to add CMake support.

I added a fully working CMake script to compile the project
automatically and to ease the use with other projects.

I also added the tinyBLAKE3 to the mix to replace the current SHA-512
in the near future.